Add a basic market implementation
Each good will have a market that responds to changes in supply based on its
max price and elasticity of demand. Elasticity is a correlation between supply
and price. The max price determins the current max price at which the demanded
quantity is 0. This price is the y intercept of the demand curve and should
change based on changes in demand.